Took all the comments in account and fixed them.

+++ HollowEngine/HollowEngine/src/Hollow/Events/EventData.h
@@ -6,7 +6,7 @@ namespace Hollow {
 		GameEvent(std::string name) {
 			mEventType = name;
 		}
-		Uint32 mTimer = 0;
+		float mTimer = 0.0f;
 		std::string mEventType;
 	};
